requirements.txt
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
cffi==1.14.0
click==7.1.1
constantly==15.1.0
crochet==1.12.0
cryptography==2.8
cssselect==1.1.0
Django==3.0.4
Flask==1.1.1
hyperlink==19.0.0
idna==2.9
image==1.5.28
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
parsel==1.5.2
Pillow==7.0.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytesseract==0.3.3
pytz==2019.3
queuelib==1.5.0
Scrapy==2.0.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
tesseract==0.1.3
Twisted==20.3.0
w3lib==1.21.0
Werkzeug==1.0.0
wrapt==1.12.1
zope.interface==5.0.1
